modified feature/doc branch
