+++ queriesdb.py
@@ -1,5 +1,10 @@
 import sqlite3
 import json
+import logging
+
+# Set up logging
+logging.basicConfig(level=logging.DEBUG)
+logger = logging.getLogger(__name__)
 
 conn = sqlite3.connect('document_database.db')
 conn.row_factory = sqlite3.Row
@@ -8,13 +13,13 @@ cursor.execute("SELECT * FROM documents;")
 rows = cursor.fetchall()
 
 for row in rows:
-    print(f"ID: {row['id']}")
-    print(f"Filename: {row['filename']}")
-    print(f"Filetype: {row['filetype']}")
-    print(f"Filesize: {row['filesize']} bytes")
-    print(f"Upload Date: {row['upload_date']}")
+    logger.info(f"ID: {row['id']}")
+    logger.info(f"Filename: {row['filename']}")
+    logger.info(f"Filetype: {row['filetype']}")
+    logger.info(f"Filesize: {row['filesize']} bytes")
+    logger.info(f"Upload Date: {row['upload_date']}")
     metadata = json.loads(row['metadata'])
-    print(f"Metadata: {metadata}")
-    print("-" * 50)
+    logger.info(f"Metadata: {metadata}")
+    logger.info("-" * 50)
 
 conn.close()
